project/requirements.txt
websockets==12.0
aiofiles==23.1.0
celery==5.2.7
fastapi==0.95.0
flower==1.2.0
Jinja2==3.1.2
pytest==7.2.2
redis==4.5.4
requests==2.28.2
uvicorn==0.21.1
httpx==0.23.3
scikit-learn==1.3.2
nltk==3.8.1
mysql-connector-python==8.2.0

python-multipart==0.0.9
pytesseract==0.3.10
pillow==10.3.0
pdf2image==1.17.0
pypdf==4.2.0
